backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.8",
    "httpx==0.27.2",
    "itsdangerous==2.2.0",
    "jinja2==3.1.4",
    "pillow==10.4.0",
    "python-dotenv==1.0.1",
    "sqladmin==0.20.1",
    "sqlalchemy==2.0.30",
    "uvicorn[standard]==0.34.0",
]
